% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoIndex.R
\name{Eco.Index}
\alias{Eco.Index}
\title{Calculates the mean ecological index for the whole pollen assemblage}
\usage{
Eco.Index(pollen_counts, pollen_plants_eco, perm = 100)
}
\arguments{
\item{pollen_counts}{data frame containing following columns: Entity, Sample, Varname, Count
\describe{
\item{Entity}{integer referring to site}
\item{Sample}{number referring  to depth or age}
\item{Varname}{character referring to pollen taxon name}
\item{Count}{number referring to pollen count}
}}

\item{pollen_plants_eco}{data frame containing the following columns: Species, Varname, EcoIndex, FreqVeg
\describe{
\item{Species}{character referring to plant species name}
\item{Varname}{character referring to pollen taxon name}
\item{EcoIndex}{number referring to ecological index}
\item{FreqVeg}{number referring to frequency of the species in vegetation (use 1 if you do not wish to apply the weighting)}
}}

\item{perm}{integer number of permutations (default perm = 100)}
}
\value{
Returns a data frame with mean EcoIndex and Standard Deviation for each sample.
}
\description{
The function requires a matrix of pollen counts and a matrix of pollen taxa linked with plant species associated with ecological index
and frequency of the species in the present-day vegetation. The function will assign plant species to each pollen grain (with permutations that can be set) and based on the species'
frequency in the vegetation (FreqVeg). Then it calculates the mean for each pollen assemblage including standard deviation.
See Kuneš et al. 2019 for details.
}
\note{
Citation of this programme: Kuneš, P., Abraham, V., & Herben, T. (2019). Changing disturbance-diversity relationships in temperate ecosystems over the past 12000 years. Journal of Ecology, 107(4), 1678–1688. doi: 10.1111/1365-2745.13136
}
\examples{

#CALCULATE DF FOR THE FIRST 3 SAMPLES OF EXAMPLE DATA

Eco.Index(PC_Prasilske[1:102,], pollen_plant_dist_2019)

}
\references{
Kuneš, P., Abraham, V., & Herben, T. (2019). Changing disturbance-diversity relationships in temperate ecosystems over the past 12000 years. Journal of Ecology, 107(4), 1678–1688. doi: 10.1111/1365-2745.13136
}
