% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PC_Prasilske}
\alias{PC_Prasilske}
\title{Pollen counts from Prášilské jezero in Šumava Mts., Czech Republic}
\format{
A data frame containing following columns: Entity, Sample, Varname, Count
\describe{
\item{Entity}{integer referring to site}
\item{Sample}{number referring  to depth or age}
\item{Varname}{character referring to pollen taxon name}
\item{Count}{number referring to pollen count}
}
}
\source{
Carter, V.A., Chiverrell, R.C., Clear, J.L., Kuosmanen, N., Moravcová, A., Svoboda, M., Svobodová-Svitavská, H., Leeuwen, V., Van Leeuwen, J., van der Knaap, W.O., & Kuneš, P. 2018. Quantitative palynology informing conservation ecology in the Bohemian/Bavarian Forests of Central Europe. Frontiers in Plant Science 8: 1–14.

Carter, V.A., Moravcová, A., Chiverrell, R.C., Clear, J.L., Finsinger, W., Dreslerová, D., Halsall, K., & Kuneš, P. 2018. Holocene-scale fire dynamics of central European temperate spruce-beech forests. Quaternary Science Reviews 191: 15–30.
}
\usage{
PC_Prasilske
}
\description{
Pollen counts from Prášilské jezero in Šumava Mts., Czech Republic
}
\examples{
data(PC_Prasilske)
\donttest{
 PC_Prasilske
}
}
\keyword{datasets}
