% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollen_plant_dist_2019}
\alias{pollen_plant_dist_2019}
\title{Data frame containing list of plant species with appropriate pollen type, disturbance frequency index and frequency of the species in the vegetation.}
\format{
A data frame with four columns:
\describe{
\item{Species}{character referring to plant species name}
\item{Varname}{character referring to pollen taxon name according to Beug (2004)}
\item{DF}{number referring to whole-community disturbance frequency index according to Herben et al. (2016)}
\item{FreqVeg}{number referring to frequency of the species in vegetation}
}
}
\source{
Kuneš, P., Abraham, V., & Herben, T. (2019). Changing disturbance-diversity relationships in temperate ecosystems over the past 12000 years. Journal of Ecology, 107(4), 1678–1688. https://doi.org/10.1111/1365-2745.13136
}
\usage{
pollen_plant_dist_2019
}
\description{
Data frame containing list of plant species with appropriate pollen type, disturbance frequency index and frequency of the species in the vegetation.
}
\examples{
data(pollen_plant_dist_2019)
\donttest{
 pollen_plant_dist_2019
}
}
\references{
Beug, H. J. (2004). Leitfaden der Pollenbestimmung für Mitteleuropa und angrenzende Gebiete. Verlag Dr. Friedrich Pfeil.

Herben, T., Chytrý, M., & Klimešová, J. (2016). A quest for species-level indicator values for disturbance. Journal of Vegetation Science, 27(3), 628–636. https://doi.org/10.1111/jvs.12384
}
\keyword{datasets}
