% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tilia_functions.R
\name{tilia_functions}
\alias{tilia_functions}
\alias{tilia_trans}
\alias{tilia_select}
\alias{tilia_REVEALS}
\alias{tilia_percent}
\alias{tilia_PAR}
\alias{tilia_CHAR}
\alias{tilia_ecog}
\alias{tilia_taxa}
\alias{tilia_EPD}
\alias{tilia_divers}
\alias{tilia_check_rarefy}
\alias{tilia_parchp}
\alias{tilia_plot_CPA}
\alias{tilia_merge}
\alias{neotoma_tilia}
\title{Functions to convert Tilia format into list of data and metadata and further formats for different analyses}
\usage{
tilia_trans(tilia_file)

tilia_select(tilia_t, agemin, agemax)

tilia_REVEALS(tilia_t, mean.age = list(), SD.age = list(), taxalist)

tilia_percent(tilia_t, groups_sum, groups, elements)

tilia_PAR(
  tilia_t,
  groups = list("TRSH", "UPHE"),
  elements = list("pollen", "spores"),
  piwot.wider = TRUE
)

tilia_CHAR(
  tilia_t,
  groups = list("CHAR"),
  elements = list("area"),
  unit = list("mm^2"),
  piwot.wider = TRUE
)

tilia_ecog(tilia_t, groups, elements)

tilia_taxa(tilia_wide)

tilia_EPD(tilia_t, epd_p_vars)

tilia_divers(
  tilia_t,
  PAR = NULL,
  index = "rarefy",
  lowsum = 300,
  groups = list("TRSH", "UPHE"),
  elements = list("pollen", "spore")
)

tilia_check_rarefy(
  tilia_t,
  PAR = null,
  groups = list("TRSH", "UPHE"),
  elements = list("pollen", "spore")
)

tilia_parchp(
  tilia_PAR_wide,
  taxon,
  method = "PELT",
  penalty = "BIC",
  minseglen,
  all.samples = F
)

tilia_plot_CPA(tilia_t, taxon, method, penalty, minseglen)

tilia_merge(
  tilia_CHAR,
  tilia_percent,
  tilia_divers,
  tilia_CPA,
  taxa = list("Pinus", "Picea", "Fagus", "Abies"),
  CHARunits = "mm^2"
)

neotoma_tilia(n_samples)
}
\arguments{
\item{tilia_file}{data frame in Tilia-format spreadsheet containing samples as columns and taxa as rows}

\item{tilia_t}{output object from tilia_trans() containing list of two data frames with data and metadata}

\item{mean.age}{list of mean ages of time slices for}

\item{age_min}{minimum age for the selection}

\item{age_max}{maximum age for the selection}
}
\value{
Returns list of two data frames containing data and metadata.
}
\description{
Functions to convert Tilia format into list of data and metadata and further formats for different analyses
}
\section{Functions}{
\itemize{
\item \code{tilia_trans()}: Convert Tilia format into list data frames with data and metadata

\item \code{tilia_select()}: Select data based on age

\item \code{tilia_REVEALS()}: Select and aggregate samples based on time slices for REVEALS analysis in LRA package

\item \code{tilia_percent()}: Function to calculate percentages for selected groups

\item \code{tilia_PAR()}: Function to calculate pollen acummulation rates

\item \code{tilia_CHAR()}: Function to calculate charcoal accumulation rate

\item \code{tilia_ecog()}: Function to extract ecological groups

\item \code{tilia_taxa()}: Function to define taxa to plot - order taxa based on sum of values

\item \code{tilia_EPD()}: Harmonize pollen taxonomy with EPD accepted taxa

\item \code{tilia_divers()}: Calculate diversity indexes

\item \code{tilia_check_rarefy()}: Check for the minimum sums for rarefaction

\item \code{tilia_parchp()}: Calculate changepoint analysis

\item \code{tilia_plot_CPA()}: Plot changepoint analysis

\item \code{tilia_merge()}: Merge different proxies for ordination

\item \code{neotoma_tilia()}: Converts Neotoma sample() data into tilia_t

}}
\examples{
#Transform data from Tilia format
tilia_trans(Tilia_Prasilske)
}
