% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tilia_Prasilske}
\alias{Tilia_Prasilske}
\title{Pollen counts from Prášilské jezero in Šumava Mts., Czech Republic, with metada in Tilia format spreadsheet}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 126 rows and 203 columns.
}
\source{
Carter, V.A., Chiverrell, R.C., Clear, J.L., Kuosmanen, N., Moravcová, A., Svoboda, M., Svobodová-Svitavská, H., Leeuwen, V., Van Leeuwen, J., van der Knaap, W.O., & Kuneš, P. 2018. Quantitative palynology informing conservation ecology in the Bohemian/Bavarian Forests of Central Europe. Frontiers in Plant Science 8: 1–14.

Carter, V.A., Moravcová, A., Chiverrell, R.C., Clear, J.L., Finsinger, W., Dreslerová, D., Halsall, K., & Kuneš, P. 2018. Holocene-scale fire dynamics of central European temperate spruce-beech forests. Quaternary Science Reviews 191: 15–30.
}
\usage{
Tilia_Prasilske
}
\description{
Pollen counts from Prášilské jezero in Šumava Mts., Czech Republic, with metada in Tilia format spreadsheet
}
\examples{
data(Tilia_Prasilske)
\donttest{
 Tilia_Prasilske
}
}
\keyword{datasets}
